refactor: centralize agent chat session logic
